organize context files in separate directory
