+++ ui/src/contexts/NavbarContext.js
